refactor: 유틸성 클래스 분리, 계층 구조 생성, 캡슐화 적용을 고려한 접근제어자 변경
